refactor: clean up utils.ts and types.ts, update tsconfig, update E2E test imports
